requirements.txt
zhipuai>=2.0.0
aiohttp>=3.8.0
Pillow>=9.0.0
tqdm>=4.64.0
pyinstaller>=5.0.0
